+++ tests/good/stdlib/cat.c
@@ -15,8 +15,8 @@ extern struct file *stdin, *stdout, *stderr;
 extern struct file *fopen(const char *filename, const char *mode);
 extern int fclose(struct file *stream);
 extern int ferror(struct file *stream);
-extern int fread(void *ptr, int size, int count, struct file *stream);
-extern int fwrite(const void *ptr, int size, int count, struct file *stream);
+extern long fread(void *ptr, long size, long count, struct file *stream);
+extern long fwrite(const void *ptr, long size, long count, struct file *stream);
 extern int fprintf(struct file *stream, const char *format, ...);
 
 extern void perror(const char *msg);
@@ -35,7 +35,7 @@ static void file_error(const char *filename) {
 }
 
 static void cat(const char *filename, struct file *src) {
-    int n;
+    long n;
     while ((n = fread(rdbuf, 1, BUF_SIZE, src)) > 0) {
         if (fwrite(rdbuf, 1, n, stdout) < n) {
             file_error(filename);
